feat: uniswap wallet on wcv2
